requirements.txt
werkzeug==2.0.3
Flask==2.1.2
gunicorn==20.1.0
torch
torchvision
Pillow==8.4.0
requests==2.26.0
numpy
